ZED-33 deduplicate test logic
