+++ backend/src/domain/repositories/ingredients/mod.rs
@@ -1,7 +1,9 @@
 pub mod errors;
+pub mod in_memory;
 pub mod postgres;
 
-pub mod in_memory;
+#[cfg(test)]
+pub mod __test__;
 
 #[cfg(test)]
 pub use in_memory::InMemoryIngredientRepository;
