+++ backend/src/domain/repositories/recipe/postgres/tests.rs
@@ -1,156 +1,66 @@
-use std::{collections::BTreeMap, time::Duration};
-
-use crate::{
-    domain::{
-        entities::recipe::ServingsType,
-        repositories::ingredients::{postgres::PostgresIngredientRepository, IngredientRepository},
-    },
-    test_utils::{recipe_changeset, recipe_fixture},
+use crate::domain::repositories::{
+    ingredients::postgres::PostgresIngredientRepository, recipe::__test__,
 };
-use pretty_assertions::assert_eq;
 
 use super::*;
-
-async fn insert_all_ingredients(ingredient_repo: impl IngredientRepository, recipe: &Recipe) {
-    join_all(
-        recipe
-            .ingredients
-            .as_ref()
-            .iter()
-            .map(|i| async { ingredient_repo.insert(i.ingredient.clone()).await.unwrap() }),
-    )
-    .await;
-}
 
 #[sqlx::test]
 async fn creating_recipe_works(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
     let ingredient_repo = PostgresIngredientRepository::new(pool);
 
-    let recipe = recipe_fixture();
-
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-
-    let result = repo.insert(recipe.clone()).await.unwrap();
-    assert_eq!(recipe, result);
+    __test__::creating_recipe_works(repo, ingredient_repo).await
 }
 
 #[sqlx::test]
 async fn inserting_recipe_with_same_id_fails(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
     let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
-
-    let recipe = recipe_fixture();
-
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-
-    repo.insert(recipe.clone()).await.unwrap();
-
-    let error = repo.insert(recipe.clone()).await.unwrap_err();
-
-    assert!(matches!(error, InsertRecipeError::Conflict(a) if a == "recipe id"));
+    __test__::inserting_recipe_with_same_id_fails(repo, ingredient_repo).await
 }
 
 #[sqlx::test]
 async fn getting_recipe_by_id_works(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
     let ingredient_repo = PostgresIngredientRepository::new(pool);
-
-    let recipe = recipe_fixture();
-
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-
-    repo.insert(recipe.clone()).await.unwrap();
-
-    let result = repo.get_by_id(&recipe.id).await.unwrap();
-
-    assert_eq!(result, recipe);
+    __test__::getting_recipe_by_id_works(repo, ingredient_repo).await
 }
 
 #[sqlx::test]
-async fn getting_a_nonexistent_recipe_errors(pool: PgPool) {
+async fn sqlx_getting_a_nonexistent_recipe_errors(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
-    let error = repo.get_by_id(&Uuid::nil()).await.unwrap_err();
-
-    assert!(matches!(error, GetRecipeByIdError::NotFound(id) if id == Uuid::nil()));
+    __test__::getting_a_nonexistent_recipe_errors(repo).await
 }
 
 #[sqlx::test]
 async fn deleting_a_recipe_succeeds(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
     let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
-    let recipe = recipe_fixture();
-
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-
-    let result = repo.insert(recipe.clone()).await.unwrap();
-
-    repo.delete(&result.id).await.unwrap();
+    __test__::deleting_a_recipe_succeeds(repo, ingredient_repo).await
 }
 
 #[sqlx::test]
 async fn deleting_a_nonexistent_recipe_fails(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool.clone());
-    let recipe = recipe_fixture();
-    let result = repo.delete(&recipe.id).await.unwrap_err();
-
-    assert!(matches!(result, DeleteRecipeError::NotFound(id) if id == recipe.id))
+    __test__::deleting_a_nonexistent_recipe_fails(repo).await
 }
 
 #[sqlx::test]
 async fn updating_a_recipe_succeeds(pool: PgPool) {
     let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
     let repo = PostgresRecipeRepository::new(pool);
-
-    let recipe = recipe_fixture();
-    let changeset = recipe_changeset();
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-
-    let result = repo.insert(recipe.clone()).await.unwrap();
-    repo.update(&result.id, changeset).await.unwrap();
-    let result = repo.get_by_id(&result.id).await.unwrap();
-
-    assert_eq!(
-        result,
-        Recipe {
-            name: "WE UPDATED THIS THING".to_string(),
-            description: "WE UPDATED THAT THING".to_string(),
-            steps: vec!["WE UPDATED ANOTHER THING".to_string()]
-                .try_into()
-                .unwrap(),
-            time: BTreeMap::from([("Prep time".to_string(), Duration::from_secs(60))]),
-            servings: ServingsType::Exact(4),
-            ..recipe
-        }
-    );
+    __test__::updating_a_recipe_succeeds(repo, ingredient_repo).await
 }
 
 #[sqlx::test]
 async fn updating_a_nonexistent_recipe_fails(pool: PgPool) {
     let repo = PostgresRecipeRepository::new(pool);
-    let recipe = recipe_fixture();
-    let changeset = RecipeChangeset {
-        name: Some("WE UPDATED THIS THING".to_string()),
-        ..Default::default()
-    };
-    let result = repo.update(&recipe.id, changeset).await.unwrap_err();
-
-    assert!(
-        matches!(result, UpdateRecipeError::Get(GetRecipeByIdError::NotFound(id)) if id == recipe.id)
-    )
+    __test__::updating_a_nonexistent_recipe_fails(repo).await
 }
 
 #[sqlx::test]
 async fn updating_a_recipe_with_empty_changeset_does_nothing(pool: PgPool) {
     let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
     let repo = PostgresRecipeRepository::new(pool);
-    let recipe = recipe_fixture();
-    let changeset = RecipeChangeset {
-        ..Default::default()
-    };
-    insert_all_ingredients(ingredient_repo, &recipe).await;
-    let result = repo.insert(recipe.clone()).await.unwrap();
-    let result = repo.update(&result.id, changeset).await.unwrap();
-
-    assert_eq!(recipe, result);
+    __test__::updating_a_recipe_with_empty_changeset_does_nothing(repo, ingredient_repo).await
 }
